Move client services into dedicated folder
