+++ WanderLost/WanderLost/Client/Pages/Notifications.razor.cs
@@ -1,5 +1,5 @@
-﻿using Blazored.LocalStorage;
-using Microsoft.AspNetCore.Components;
+﻿using Microsoft.AspNetCore.Components;
+using WanderLost.Client.Services;
 using WanderLost.Shared.Data;
 
 namespace WanderLost.Client.Pages
