+++ WanderLost/WanderLost/Client/_Imports.razor
@@ -8,6 +8,7 @@
 @using Microsoft.JSInterop
 @using WanderLost.Client
 @using WanderLost.Client.Components
+@using WanderLost.Client.Services
 @using WanderLost.Client.Shared
 @using WanderLost.Shared
 @using WanderLost.Shared.Data
